% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RockandRose-package.R
\docType{package}
\name{RockandRose-package}
\alias{RockandRose}
\alias{RockandRose-package}
\title{RockandRose: RockandRose}
\description{
Explore financial data for rock and rose landscapes llc.
}
\author{
\strong{Maintainer}: Jacob Braslaw \email{jacobbraslaw@ucsb.edu}

}
\keyword{internal}
