% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RockandRose-package.R
\docType{package}
\name{RockandRose-package}
\alias{RockandRose}
\alias{RockandRose-package}
\title{RockandRose: An Amazing Shiny App}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: firstname lastname \email{your@email.com}

}
\keyword{internal}
